[fix]: rename database class to attend propject design pattern
