+++ src/repositories/AdminRepository.js
@@ -1,4 +1,4 @@
-const database = require('../database/Database.js');
+const database = require('../database/database.js');
 const Admin = require('../model/AdminModel.js');
 
 // class AdminRepository{
