+++ src/repositories/OrganizationRepository.js
@@ -1,4 +1,4 @@
-const database = require('../database/Database.js');
+const database = require('../database/database.js');
 const Organization = require('../model/OrganizationModel.js');
 
 class OrganizationRepository {
